requirements.txt
torch>=1.10.0
torchvision>=0.11.1
numpy>=1.21.2
opencv-python>=4.5.4
pytorch-msssim>=0.2.1
PyYAML>=5.4.1